+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - restructure readme:
   add a readme for each test suite
 - remove `age-test-roms/_in-progress` as these tests are not finished yet
+- add custom `manual-only/sprite_priority` screenshots (Mooneye Test Suite)
 
 ## [v4.0](https://github.com/c-sp/gameboy-test-roms/releases/tag/v4.0) *(2022-03-30)*